Refactor middleware.go: Add authRequired method
Add a new method `authRequired` to the `middleware.go` file. This method is responsible for checking the authentication token in the request header and verifying it. If the token is not valid or missing, it returns a `401 Unauthorized` response. This method is used as a middleware to protect certain routes that require authentication.