Add functionality to add a new todo
This commit adds the functionality to add a new todo item to the application. The `AddTodo` handler is implemented to handle the HTTP PUT request for adding a todo. The handler reads the request payload, inserts the todo into the database, and returns a JSON response with the newly created todo's ID and user ID. The `InsertTodo` method is added to the `DatabaseRepo` interface and implemented in the `PostgresDBRepo` struct to handle the database insertion.